UI feedback for successful promo code
Also setup submit form for production file upload. There is a bug with the
promo code success message. For some reason once we change the
message on a success, we cannot go back to the default message on a
subsequent fail even though we are trying to reload the default imported
message.